Add simple SSO authentication
If OPENID_TEAMS is set (currently ["ubuntuone-hackers"] for staging and
production), then we require the user to prove membership of at least
one of those teams before we let them use the site.  This gives us a
quick way to restrict widespread use before everything is ready.
Long-term, we expect to require GitHub credentials instead.

Almost all of this is cloned-and-hacked from javan-rhino.  The
authorisation is hooked up slightly differently, since in this case
we're using SSO as a front door rather than extracting any meaningful
information from it.