Hook up snap creation in the frontend
If LP_API_USERNAME is set, then we now create a snap based on the given
GitHub repository rather than merely verifying it.  This uses the
/launchpad/snaps API to create the snap in Launchpad, then redirects
through OpenID to authorise uploads of it to the store, then redirects
to the builds page for that repository.

If LP_API_USERNAME is not set, then we behave as before and only verify
the repository.  This is partly because it's handy in development, and
partly because we still need the verify actions to support future plans.