Create snaps with more appropriate architectures
Launchpad defaults to [amd64, i386], which is possibly not ideal for
snaps.  Let's go for [amd64, armhf] instead; we need this to be
user-selectable in the long run, but this will be better for demo
purposes.

In the process, I fixed some buggy use of nock in tests: the Launchpad
client passes parameters to POST requests using the request body, not
the query string.