Finish hooking up completeSnapAuthorization
It appears that I never actually arranged for completeSnapAuthorization
to be called on creating a new snap, so let's fix that.  The
authorization workflow goes through the server-rendered page
/login/verify after returning from SSO, so the simplest way to do things
is just to have that call completeSnapAuthorization directly rather than
via a route.  We need to add the GitHub repository URL to the parameters
passed through the OpenID dance in order for this to work.