Add routes to retrieve SSO discharge from session
Nothing stores a discharge macaroon in the session yet, but this will be
useful once we've reworked our authorisation workflow to use a
package_upload_request macaroon.

I suspect the new actions won't actually work in browsers yet since
macaroons.js needs a little bit of polyfilling, but that'll be easier to
test and fix up once a few more pieces are in place.  It's harmless for
now since the new actions aren't used yet.