Convert all uses of memcached to promises
This is much easier to use, especially in conjunction with other code
using promises.  The test stubs already used promises, but this had a
dangerous gotcha: code that expected memcached methods to return
promises would work under test but fail in production, as was the case
with the webhook implementation.  Using promises everywhere avoids this
problem.

I considered using one of the existing memcached/promise integration
libraries, but all the ones I could find depended on some specific
promise library, and it's so easy to do by hand that it wasn't worth the
hassle.