Store GitHub user information in a database
We now use the Bookshelf ORM and the Knex query builder, which gives us
a promise-based interface to multiple databases with minimal
boilerplate.

In development, we use SQLite by default, since that involves minimal
setup.  In staging and production, we use PostgreSQL, so this comes with
charm changes to ensure that we have access to a suitable database.

At the moment, user information is saved to both the session and the
database.  This is obviously duplicative, and we should perhaps consider
having just an ID in the session and fetching the rest from the database
as needed.  Similarly, perhaps some of the things we use memcached for
would make more sense in a more persistent database instead.

Part of #360.