Run database queries in transactions
At the moment we have no particular need for cleanly rolling back
transactions, but we will soon, and we might as well put good practice
in place now.

I considered using `bookshelf-transaction-manager` to avoid the
`{ transacting: trx }` boilerplate, but it ended up being much more
trouble than it was worth: it doesn't currently wrap all the Bookshelf
model/collection methods, which makes it very confusing and error-prone
to use.