Add instance option for LP credentials
BSI can be usefully configured locally to point to Launchpad instances
other than production for the purpose of testing various things: staging
mostly works (modulo a snap-proxy problem that'll be fixed soon),
dogfood works, and it's even possible to use a local development
instance.  To make such things easier, add a `--launchpad` option to
`scripts/create-launchpad-credentials`.

Production is still the right choice for BSI staging for the moment,
since Launchpad staging is reset weekly and would require some
reconfiguration to preserve BSI credentials across the reset, but it may
be worth looking into that in future.