Add baseline support for organisation repositories
This shows organisation repositories for which the user has admin
permissions in the add-repositories dialog, and shows organisation
repositories that the user can write to on the my-repositories page.
Various aspects of the UI will need to be tidied up in future PRs, but
this at least gets us started.

We now request the `read:org` scope from GitHub, which allows us to tell
which organisations a user is in.

Fixes #620; fixes #648.