Attribute organization repo activity to registrant
Keep track of who registered each repository with build.snapcraft.io in
a new Repository table, and attribute `builds_requested` and
`builds_released` activity to that user.  Catch up with existing
repositories via the `/api/launchpad/snaps/list` handler.

This isn't necessarily perfect since the publisher can be reassigned in
the store without us noticing, but this is tolerable since it's only for
metrics.

The `Repository.snapcraft_name` and `Repository.store_name` columns
aren't immediately needed for this, but will be useful for future work
such as #706.

Fixes #621.