Fix store expiry check
We switched to the forked store recently, which means that the host part
of STORE_API_URL is `dashboard.snapcraft.io`, but it turns out that the
forked store still issues macaroons whose location is
`myapps.developer.ubuntu.com`, presumably out of concerns for
compatibility.  When extracting the `expires` caveat, we therefore need
to extract one that matches the macaroon's location, rather than
assuming that the location is going to match the host that we thought we
were contacting.

Fixes #719.