Work around ols-jenkaas build-charm failures
ols-jenkaas is set up such that code running in the containers doesn't
currently have a git identity configured, so the build-charm jobs
sometimes fail depending on whether git manages to auto-detect some
semblance of a reasonable identity.  Work around this by forcing the
correct identity, although this is hardcoding production configuration
that ought to live elsewhere.