Use default branch rather than hardcoding master
The links to create/edit `snapcraft.yaml` were problematic: we haven't
necessarily fetched repository information from GitHub yet when these
are needed, and doing so was going to be quite cumbersome.  Instead, I
introduced server endpoints that fetch the current default branch from
GitHub and redirect to the appropriate URL, which is much simpler.

Fixes #497.