Check the pushed ref in the GitHub webhook handler
We noticed that the Launchpad build farm was being spammed by identical
builds of certain snaps.  On investigation, this turned out to be
because BSI's GitHub webhook handler doesn't check that the pushed
branch is the one that Launchpad is configured to build for this
repository, so dispatched builds even in the rather common case that the
push was to some other branch.  Checking the branch should mean that we
have more resources for builds that aren't redundant.