Return pending build requests from getSnapBuilds
This allows the UI to show the user that there's a pending build
request, rather than appearing unresponsive in the interval between a
set of builds being requested and the request being processed.

The database gains a new `build_request_annotation` table, allowing us
to keep track of build reasons even when the process of creating new
builds goes through a build request.

I'd like to return failed build requests as well, but there are some
significant complications involved in making that work efficiently in
conjunction with pagination, mainly on the Launchpad side.  I don't
think that needs to block the rest of this work.

Part of #556.